requirements.txt
pyyaml
httpx
uvicorn[standard]
fastapi
tiktoken
azure-identity
azure-monitor-ingestion